Make pointer validation in host_functions configurable using the execution interface.  Simplifies integration with OC.
